@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "animejs": "^3.0.1",
-    "next": "^8.1.0",
+    "next": "^9.4.4",
     "react": "^16.5.2",
     "react-dom": "^16.5.2"
   },
